feat(babel): add babel transpiler and its dependencies
Some of the EcmaScript 6 (ES6) functionalies are not supported in all browsers
one of them for example is the spread syntax ...
In order to take advantage of it, we are using the babel transpiler. Add babel and
its dependencies to the project, create a .babelrc file with the babel properties,
update the webpack basse config file with the properties and use the babel eslint parser.