feat(lazy-loading): loading components dynamically when needed
often times we do not need all components to be loaded when the application starts.
we only need specific set of components at any point of time based on the pages and data
users want to view. Lets take advantage of the async loading of vue, code splitting of webpack
combined with dynamic import syntax of babel to enable lazy loading in our application.

this lazy loading is extremely useful when the size of the application increases.