feat(initial-state): add asyncData, initial state to the application
with server side rendering even though the code is generated and sent to the UI, there is still
an inconsistency between client side and server side to check if both have the same dom content.
this results in following error in the browser console:

[Vue warn]: The client-side rendered virtual DOM tree is not matching server-rendered content.
This is likely caused by incorrect HTML markup, for example nesting block-level elements inside
<p>, or missing <tbody>. Bailing hydration and performing full client-side render.

In order to resolve this issue, we need to stamp the server side rendered code in the UI and make
sure that the states are in sync. this is achieved by setting the initial state and state matching